+++ banyand/internal/sidx/sidx.go
@@ -21,10 +21,8 @@ import (
 	"container/heap"
 	"context"
 	"encoding/json"
-	"fmt"
 	"math"
 	"path/filepath"
-	"slices"
 	"sort"
 	"sync"
 	"sync/atomic"
@@ -35,19 +33,15 @@ import (
 	modelv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/model/v1"
 	"github.com/apache/skywalking-banyandb/banyand/internal/storage"
 	"github.com/apache/skywalking-banyandb/banyand/protector"
-	"github.com/apache/skywalking-banyandb/pkg/encoding"
-	"github.com/apache/skywalking-banyandb/pkg/fs"
 	"github.com/apache/skywalking-banyandb/pkg/cgroups"
+	"github.com/apache/skywalking-banyandb/pkg/fs"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
 	"github.com/apache/skywalking-banyandb/pkg/pool"
 	"github.com/apache/skywalking-banyandb/pkg/run"
 	"github.com/apache/skywalking-banyandb/pkg/watcher"
 )
 
-const (
-	maxUncompressedBlockSize = 2 * 1024 * 1024
-	maxBlockLength           = 8 * 1024
-)
+const maxBlockLength = 8 * 1024
 
 // sidx implements the SIDX interface with introduction channels for async operations.
 type sidx struct {
@@ -143,42 +137,6 @@ func (s *sidx) Write(ctx context.Context, reqs []WriteRequest, partID uint64) er
 	}
 }
 
-// extractKeyRange extracts min/max key range from QueryRequest.
-func extractKeyRange(req QueryRequest) (int64, int64) {
-	minKey := int64(math.MinInt64)
-	maxKey := int64(math.MaxInt64)
-
-	if req.MinKey != nil {
-		minKey = *req.MinKey
-	}
-	if req.MaxKey != nil {
-		maxKey = *req.MaxKey
-	}
-
-	return minKey, maxKey
-}
-
-// extractOrdering extracts ordering direction from QueryRequest.
-func extractOrdering(req QueryRequest) bool {
-	if req.Order == nil {
-		return true // Default ascending
-	}
-	return req.Order.Sort != modelv1.Sort_SORT_DESC
-}
-
-// selectPartsForQuery selects relevant parts from snapshot based on key range.
-func selectPartsForQuery(snap *snapshot, minKey, maxKey int64) []*part {
-	var selectedParts []*part
-
-	for _, pw := range snap.parts {
-		if pw.isActive() && pw.overlapsKeyRange(minKey, maxKey) {
-			selectedParts = append(selectedParts, pw.p)
-		}
-	}
-
-	return selectedParts
-}
-
 // Query implements SIDX interface.
 func (s *sidx) Query(ctx context.Context, req QueryRequest) (*QueryResponse, error) {
 	if err := req.Validate(); err != nil {
@@ -260,6 +218,42 @@ func (s *sidx) Query(ctx context.Context, req QueryRequest) (*QueryResponse, err
 	}
 
 	return response, nil
+}
+
+// extractKeyRange extracts min/max key range from QueryRequest.
+func extractKeyRange(req QueryRequest) (int64, int64) {
+	minKey := int64(math.MinInt64)
+	maxKey := int64(math.MaxInt64)
+
+	if req.MinKey != nil {
+		minKey = *req.MinKey
+	}
+	if req.MaxKey != nil {
+		maxKey = *req.MaxKey
+	}
+
+	return minKey, maxKey
+}
+
+// extractOrdering extracts ordering direction from QueryRequest.
+func extractOrdering(req QueryRequest) bool {
+	if req.Order == nil {
+		return true // Default ascending
+	}
+	return req.Order.Sort != modelv1.Sort_SORT_DESC
+}
+
+// selectPartsForQuery selects relevant parts from snapshot based on key range.
+func selectPartsForQuery(snap *snapshot, minKey, maxKey int64) []*part {
+	var selectedParts []*part
+
+	for _, pw := range snap.parts {
+		if pw.isActive() && pw.overlapsKeyRange(minKey, maxKey) {
+			selectedParts = append(selectedParts, pw.p)
+		}
+	}
+
+	return selectedParts
 }
 
 // executeBlockScannerQuery coordinates the worker pool with block scanner following tsResult pattern.
@@ -409,6 +403,7 @@ func (s *sidx) Flush() error {
 
 	// Create flush introduction
 	flushIntro := generateFlusherIntroduction()
+	defer releaseFlusherIntroduction(flushIntro)
 	flushIntro.applied = make(chan struct{})
 
 	// Select memory parts to flush
@@ -423,7 +418,6 @@ func (s *sidx) Flush() error {
 	}
 
 	if len(flushIntro.flushed) == 0 {
-		releaseFlusherIntroduction(flushIntro)
 		return nil
 	}
 
@@ -432,54 +426,6 @@ func (s *sidx) Flush() error {
 
 	// Wait for flush to complete
 	<-flushIntro.applied
-	releaseFlusherIntroduction(flushIntro)
-
-	return nil
-}
-
-// Merge implements Merger interface.
-func (s *sidx) Merge(partIDs []uint64, newPartID uint64, closeCh <-chan struct{}) error {
-	// Get current snapshot
-	snap := s.currentSnapshot()
-	if snap == nil {
-		return nil
-	}
-	defer snap.decRef()
-
-	// Create merge introduction
-	mergeIntro := generateMergerIntroduction()
-	defer releaseMergerIntroduction(mergeIntro)
-	mergeIntro.applied = make(chan struct{})
-
-	// Select parts to merge
-	var partsToMerge []*partWrapper
-	for _, pw := range snap.parts {
-		if pw.isActive() && !pw.isMemPart() && slices.Contains(partIDs, pw.ID()) {
-			partsToMerge = append(partsToMerge, pw)
-		}
-	}
-
-	if len(partsToMerge) < 2 {
-		return nil
-	}
-
-	// Mark parts for merging
-	for _, pw := range partsToMerge {
-		mergeIntro.merged[pw.ID()] = struct{}{}
-	}
-
-	// Create new merged part
-	newPart, err := s.mergeParts(s.fileSystem, closeCh, partsToMerge, newPartID, s.root)
-	if err != nil {
-		return err
-	}
-	mergeIntro.newPart = newPart.p
-
-	// Send to introducer loop
-	s.mergeCh <- mergeIntro
-
-	// Wait for merge to complete
-	<-mergeIntro.applied
 
 	return nil
 }
@@ -515,183 +461,6 @@ func (s *sidx) currentSnapshot() *snapshot {
 	return nil
 }
 
-var errNoPartToMerge = fmt.Errorf("no part to merge")
-
-func (s *sidx) mergeParts(fileSystem fs.FileSystem, closeCh <-chan struct{}, parts []*partWrapper, partID uint64, root string) (*partWrapper, error) {
-	if len(parts) == 0 {
-		return nil, errNoPartToMerge
-	}
-	dstPath := partPath(root, partID)
-	var totalSize int64
-	pii := make([]*partMergeIter, 0, len(parts))
-	for i := range parts {
-		pmi := generatePartMergeIter()
-		pmi.mustInitFromPart(parts[i].p)
-		pii = append(pii, pmi)
-		totalSize += int64(parts[i].p.partMetadata.CompressedSizeBytes)
-	}
-	shouldCache := s.pm.ShouldCache(totalSize)
-	br := generateBlockReader()
-	br.init(pii)
-	bw := generateBlockWriter()
-	bw.mustInitForFilePart(fileSystem, dstPath, shouldCache)
-
-	pm, err := mergeBlocks(closeCh, bw, br)
-	releaseBlockWriter(bw)
-	releaseBlockReader(br)
-	for i := range pii {
-		releasePartMergeIter(pii[i])
-	}
-	if err != nil {
-		return nil, err
-	}
-	pm.mustWriteMetadata(fileSystem, dstPath)
-	fileSystem.SyncPath(dstPath)
-	p := mustOpenPart(dstPath, fileSystem)
-	return newPartWrapper(nil, p), nil
-}
-
-var errClosed = fmt.Errorf("the merger is closed")
-
-func mergeBlocks(closeCh <-chan struct{}, bw *blockWriter, br *blockReader) (*partMetadata, error) {
-	pendingBlockIsEmpty := true
-	pendingBlock := generateBlockPointer()
-	defer releaseBlockPointer(pendingBlock)
-	var tmpBlock, tmpBlock2 *blockPointer
-	var decoder *encoding.BytesBlockDecoder
-	getDecoder := func() *encoding.BytesBlockDecoder {
-		if decoder == nil {
-			decoder = generateColumnValuesDecoder()
-		}
-		return decoder
-	}
-	releaseDecoder := func() {
-		if decoder != nil {
-			releaseColumnValuesDecoder(decoder)
-			decoder = nil
-		}
-	}
-	for br.nextBlockMetadata() {
-		select {
-		case <-closeCh:
-			return nil, errClosed
-		default:
-		}
-		b := br.block
-
-		if pendingBlockIsEmpty {
-			br.loadBlockData(getDecoder())
-			pendingBlock.copyFrom(b)
-			pendingBlockIsEmpty = false
-			continue
-		}
-
-		if pendingBlock.bm.seriesID != b.bm.seriesID ||
-			(pendingBlock.isFull() && pendingBlock.bm.maxKey <= b.bm.minKey) {
-			bw.mustWriteBlock(pendingBlock.bm.seriesID, &pendingBlock.block)
-			releaseDecoder()
-			br.loadBlockData(getDecoder())
-			pendingBlock.copyFrom(b)
-			continue
-		}
-
-		if tmpBlock == nil {
-			tmpBlock = generateBlockPointer()
-			defer releaseBlockPointer(tmpBlock)
-		}
-		tmpBlock.reset()
-		tmpBlock.bm.seriesID = b.bm.seriesID
-		br.loadBlockData(getDecoder())
-		mergeTwoBlocks(tmpBlock, pendingBlock, b)
-		if len(tmpBlock.userKeys) <= maxBlockLength && tmpBlock.uncompressedSizeBytes() <= maxUncompressedBlockSize {
-			if len(tmpBlock.userKeys) == 0 {
-				pendingBlockIsEmpty = true
-			}
-			pendingBlock, tmpBlock = tmpBlock, pendingBlock
-			continue
-		}
-
-		if len(tmpBlock.userKeys) <= maxBlockLength {
-			bw.mustWriteBlock(tmpBlock.bm.seriesID, &tmpBlock.block)
-			pendingBlock.reset()
-			pendingBlockIsEmpty = true
-			releaseDecoder()
-			continue
-		}
-		tmpBlock.idx = maxBlockLength
-		pendingBlock.copyFrom(tmpBlock)
-		l := tmpBlock.idx
-		tmpBlock.idx = 0
-		if tmpBlock2 == nil {
-			tmpBlock2 = generateBlockPointer()
-			defer releaseBlockPointer(tmpBlock2)
-		}
-		tmpBlock2.reset()
-		tmpBlock2.append(tmpBlock, l)
-		bw.mustWriteBlock(tmpBlock.bm.seriesID, &tmpBlock2.block)
-		releaseDecoder()
-	}
-	if err := br.error(); err != nil {
-		return nil, fmt.Errorf("cannot read block to merge: %w", err)
-	}
-	if !pendingBlockIsEmpty {
-		bw.mustWriteBlock(pendingBlock.bm.seriesID, &pendingBlock.block)
-	}
-	releaseDecoder()
-	var result partMetadata
-	bw.Flush(&result)
-	return &result, nil
-}
-
-func mergeTwoBlocks(target, left, right *blockPointer) {
-	appendIfEmpty := func(ib1, ib2 *blockPointer) bool {
-		if ib1.idx >= len(ib1.userKeys) {
-			target.appendAll(ib2)
-			return true
-		}
-		return false
-	}
-
-	defer target.updateMetadata()
-
-	if left.bm.maxKey < right.bm.minKey {
-		target.appendAll(left)
-		target.appendAll(right)
-		return
-	}
-	if right.bm.maxKey < left.bm.minKey {
-		target.appendAll(right)
-		target.appendAll(left)
-		return
-	}
-	if appendIfEmpty(left, right) || appendIfEmpty(right, left) {
-		return
-	}
-
-	for {
-		i := left.idx
-		uk2 := right.userKeys[right.idx]
-		for i < len(left.userKeys) && left.userKeys[i] <= uk2 {
-			i++
-		}
-		target.append(left, i)
-		left.idx = i
-		if appendIfEmpty(left, right) {
-			return
-		}
-		left, right = right, left
-	}
-}
-
-// Helper methods for metrics.
-func (s *sidx) incTotalIntroduceLoopStarted(_ string) {
-	s.totalIntroduceLoopStarted.Add(1)
-}
-
-func (s *sidx) incTotalIntroduceLoopFinished(_ string) {
-	s.totalIntroduceLoopFinished.Add(1)
-}
-
 // persistSnapshot persists the snapshot to disk.
 func (s *sidx) persistSnapshot(snapshot *snapshot) {
 	var partNames []string
@@ -721,20 +490,13 @@ func (s *sidx) mustWriteSnapshot(snapshot uint64, partNames []string) {
 	}
 }
 
-// Lock/Unlock methods for introducer loop.
-func (s *sidx) Lock() {
-	s.mu.Lock()
+// Helper methods for metrics.
+func (s *sidx) incTotalIntroduceLoopStarted(_ string) {
+	s.totalIntroduceLoopStarted.Add(1)
 }
 
-func (s *sidx) Unlock() {
-	s.mu.Unlock()
-}
-
-// emptyQueryResult represents an empty query result.
-type emptyQueryResult struct{}
-
-func (e *emptyQueryResult) Pull() *QueryResponse {
-	return nil
+func (s *sidx) incTotalIntroduceLoopFinished(_ string) {
+	s.totalIntroduceLoopFinished.Add(1)
 }
 
 // blockCursor represents a cursor for iterating through a loaded block, similar to query_by_ts.go.
